split out infos from state
