+++ lib/elixir_sense/core/compiler/state.ex
@@ -3,8 +3,19 @@ defmodule ElixirSense.Core.State do
   Core State
   """
 
-  alias ElixirSense.Core.Introspection
   alias ElixirSense.Core.BuiltinFunctions
+  alias ElixirSense.Core.State.Env
+
+  alias ElixirSense.Core.State.{
+    CallInfo,
+    StructInfo,
+    ModFunInfo,
+    SpecInfo,
+    TypeInfo,
+    VarInfo,
+    AttributeInfo
+  }
+
   require Logger
 
   @type fun_arity :: {atom, non_neg_integer}
@@ -108,250 +119,6 @@ defmodule ElixirSense.Core.State do
             ex_unit_describe: nil,
             attribute_store: %{},
             cursor_position: nil
-
-  defmodule Env do
-    @moduledoc """
-    Line environment
-    """
-
-    @type t :: %Env{
-            functions: [{module, [{atom, arity}]}],
-            macros: [{module, [{atom, arity}]}],
-            requires: list(module),
-            aliases: list(ElixirSense.Core.State.alias_t()),
-            macro_aliases: [{module, {term, module}}],
-            context: nil | :match | :guard,
-            module: nil | module,
-            function: nil | {atom, arity},
-            protocol: nil | ElixirSense.Core.State.protocol_t(),
-            versioned_vars: %{optional({atom, atom}) => non_neg_integer},
-            vars: list(ElixirSense.Core.State.VarInfo.t()),
-            attributes: list(ElixirSense.Core.State.AttributeInfo.t()),
-            behaviours: list(module),
-            context_modules: list(module),
-            typespec: nil | {atom, arity},
-            scope_id: nil | ElixirSense.Core.State.scope_id_t()
-          }
-    defstruct functions: [],
-              macros: [],
-              requires: [],
-              aliases: [],
-              macro_aliases: [],
-              # NOTE for protocol implementation this will be the first variant
-              module: nil,
-              function: nil,
-              # NOTE for protocol implementation this will be the first variant
-              protocol: nil,
-              versioned_vars: %{},
-              vars: [],
-              attributes: [],
-              behaviours: [],
-              context_modules: [],
-              context: nil,
-              typespec: nil,
-              scope_id: nil
-
-    def to_macro_env(%__MODULE__{} = env, file \\ "nofile", line \\ 1) do
-      # we omit lexical_tracker and tracers
-      %Macro.Env{
-        line: line,
-        file: file,
-        context: env.context,
-        module: env.module,
-        function: env.function,
-        context_modules: env.context_modules,
-        macros: env.macros,
-        functions: env.functions,
-        requires: env.requires,
-        aliases: env.aliases,
-        macro_aliases: env.macro_aliases,
-        versioned_vars: env.versioned_vars
-      }
-    end
-
-    def update_from_macro_env(%__MODULE__{} = env, macro_env = %Macro.Env{}) do
-      # we omit lexical_tracker and tracers
-      %__MODULE__{
-        env
-        | context: macro_env.context,
-          module: macro_env.module,
-          function: macro_env.function,
-          context_modules: macro_env.context_modules,
-          macros: macro_env.macros,
-          functions: macro_env.functions,
-          requires: macro_env.requires,
-          aliases: macro_env.aliases,
-          macro_aliases: macro_env.macro_aliases,
-          versioned_vars: macro_env.versioned_vars
-      }
-    end
-  end
-
-  defmodule VarInfo do
-    @moduledoc """
-    Variable info
-    """
-
-    @type t :: %VarInfo{
-            name: atom,
-            positions: list(ElixirSense.Core.State.position_t()),
-            scope_id: nil | ElixirSense.Core.State.scope_id_t(),
-            version: non_neg_integer(),
-            type: ElixirSense.Core.State.var_type()
-          }
-    defstruct name: nil,
-              positions: [],
-              scope_id: nil,
-              version: 0,
-              type: nil
-  end
-
-  defmodule TypeInfo do
-    @moduledoc """
-    Type definition info
-    """
-    @type t :: %TypeInfo{
-            name: atom,
-            args: list(list(String.t())),
-            specs: [String.t()],
-            kind: :type | :typep | :opaque,
-            positions: [ElixirSense.Core.State.position_t()],
-            end_positions: [ElixirSense.Core.State.position_t() | nil],
-            doc: String.t(),
-            meta: map(),
-            generated: list(boolean)
-          }
-    defstruct name: nil,
-              args: [],
-              specs: [],
-              kind: :type,
-              positions: [],
-              end_positions: [],
-              generated: [],
-              doc: "",
-              meta: %{}
-  end
-
-  defmodule SpecInfo do
-    @moduledoc """
-    Type definition info
-    """
-    @type t :: %SpecInfo{
-            name: atom,
-            args: list(list(String.t())),
-            specs: [String.t()],
-            kind: :spec | :callback | :macrocallback,
-            positions: [ElixirSense.Core.State.position_t()],
-            end_positions: [ElixirSense.Core.State.position_t() | nil],
-            doc: String.t(),
-            meta: map(),
-            generated: list(boolean)
-          }
-    defstruct name: nil,
-              args: [],
-              specs: [],
-              kind: :spec,
-              positions: [],
-              end_positions: [],
-              generated: [],
-              doc: "",
-              meta: %{}
-  end
-
-  defmodule StructInfo do
-    @moduledoc """
-    Structure definition info
-    """
-    @type field_t :: {atom, any}
-    @type t :: %StructInfo{
-            type: :defstruct | :defexception,
-            fields: list(field_t)
-          }
-    defstruct type: :defstruct, fields: []
-  end
-
-  defmodule AttributeInfo do
-    @moduledoc """
-    Variable info
-    """
-    @type t :: %AttributeInfo{
-            name: atom,
-            positions: list(ElixirSense.Core.State.position_t()),
-            type: ElixirSense.Core.State.var_type()
-          }
-    defstruct name: nil, positions: [], type: nil
-  end
-
-  defmodule CallInfo do
-    @moduledoc """
-    Function call info
-    """
-    @type t :: %CallInfo{
-            arity: non_neg_integer,
-            position: ElixirSense.Core.State.position_t(),
-            func: atom,
-            mod: module | {:attribute, atom}
-          }
-    defstruct arity: 0,
-              position: {1, 1},
-              func: nil,
-              mod: Elixir
-  end
-
-  defmodule ModFunInfo do
-    @moduledoc """
-    Module or function info
-    """
-
-    @type t :: %ModFunInfo{
-            params: list(list(term)),
-            positions: list(ElixirSense.Core.State.position_t()),
-            end_positions: list(ElixirSense.Core.State.position_t() | nil),
-            target: nil | {module, atom},
-            overridable: false | {true, module},
-            generated: list(boolean),
-            doc: String.t(),
-            meta: map(),
-            # TODO defmodule defprotocol defimpl?
-            type:
-              :def
-              | :defp
-              | :defmacro
-              | :defmacrop
-              | :defdelegate
-              | :defguard
-              | :defguardp
-              | :defmodule
-          }
-
-    defstruct params: [],
-              positions: [],
-              end_positions: [],
-              target: nil,
-              type: nil,
-              generated: [],
-              overridable: false,
-              doc: "",
-              meta: %{}
-
-    def get_arities(%ModFunInfo{params: params_variants}) do
-      params_variants
-      |> Enum.map(fn params ->
-        {length(params), Introspection.count_defaults(params)}
-      end)
-    end
-
-    def get_category(%ModFunInfo{type: type})
-        when type in [:defmacro, :defmacrop, :defguard, :defguardp],
-        do: :macro
-
-    def get_category(%ModFunInfo{type: type}) when type in [:def, :defp, :defdelegate],
-      do: :function
-
-    def get_category(%ModFunInfo{}), do: :module
-
-    def private?(%ModFunInfo{type: type}), do: type in [:defp, :defmacrop, :defguardp]
-  end
 
   defp get_current_env(%__MODULE__{} = state, macro_env) do
     current_attributes = state |> get_current_attributes()
